+++ packages/react-heartwood-components/src/components/Forms/Forms-story.js
@@ -9,6 +9,7 @@ import {
 	number,
 	select
 } from '@storybook/addon-knobs/react'
+import moment from 'moment'
 import Button from '../Button/Button'
 import {
 	Autosuggest,
@@ -256,6 +257,7 @@ stories
 		<Fragment>
 			<Select
 				label={text('label', 'Country')}
+				placeholder={text('placeholder', 'Select something...')}
 				id={text('id', 'country')}
 				options={object('options', {
 					us: 'United States',
@@ -280,6 +282,7 @@ stories
 			onSelectDateRange={({ startDate, endDate }) =>
 				console.log(startDate, endDate)
 			}
+			isDayBlocked={day => day.isBefore(moment().startOf('day'))}
 		/>
 	))
 	.add('Stars', () => <Stars />)