+++ packages/react-heartwood-components/src/components/RecordSelectionList/RecordSelectionList.js
@@ -23,6 +23,11 @@ type RecordSelectionListProps = {|
 	/** Required method to render a record into a node */
 	renderRecord: any => Node,
 
+	/** Get a unique ID for a record; given that data may be shaped in an unpredictable manner,
+	 * you must implement this at each usage.
+	 */
+	getRecordId: Record => string,
+
 	/** Load records for the offset/limit provided */
 	loadRecords: ({|
 		offset: number,
@@ -51,6 +56,9 @@ type RecordSelectionListProps = {|
 
 	/** Optionally provide a placeholder to the search input */
 	searchPlaceholder?: string,
+
+	/** Set to false to hide "# selected" text - defaults to true */
+	showSelectedCount?: boolean,
 
 	/** Callback for selection of a record */
 	onSelect?: (RecordId, Record) => void,
@@ -84,6 +92,12 @@ export default class RecordSelectionList extends Component<
 
 	constructor(props: RecordSelectionListProps) {
 		super(props)
+
+		if (!props.getRecordId) {
+			throw new Error(
+				"RecordSelectionList: `getRecordId` must be provided to determine a record's unique identifier. (record => string)"
+			)
+		}
 
 		this.state = { loadedRecords: [], isLoading: false, search: '' }
 	}
@@ -131,6 +145,7 @@ export default class RecordSelectionList extends Component<
 			selectedIds,
 			unselectableIds,
 			renderRecord,
+			getRecordId,
 			canSelect,
 			canRemove,
 			onSelect,
@@ -158,12 +173,15 @@ export default class RecordSelectionList extends Component<
 							<SelectionComponent
 								className="record-selection__record-select"
 								onChange={() => {
-									onSelect(record.id, record)
+									onSelect(getRecordId(record), record)
 								}}
 								disabled={
-									unselectableIds && unselectableIds.indexOf(record.id) >= 0
+									unselectableIds &&
+									unselectableIds.indexOf(getRecordId(record)) >= 0
 								}
-								checked={selectedIds && selectedIds.indexOf(record.id) >= 0}
+								checked={
+									selectedIds && selectedIds.indexOf(getRecordId(record)) >= 0
+								}
 							/>
 						)}
 
@@ -180,11 +198,12 @@ export default class RecordSelectionList extends Component<
 								onClick={() => {
 									this.setState({
 										loadedRecords: loadedRecords.filter(
-											loadedRecord => loadedRecord.id !== record.id
+											loadedRecord =>
+												getRecordId(loadedRecord) !== getRecordId(record)
 										)
 									})
 
-									onRemove(record.id, record)
+									onRemove(getRecordId(record), record)
 								}}
 							/>
 						)}
@@ -259,7 +278,8 @@ export default class RecordSelectionList extends Component<
 			selectedIds = [],
 			totalRecordCount,
 			canSearch,
-			searchPlaceholder
+			searchPlaceholder,
+			showSelectedCount
 		} = this.props
 		const { loadedRecords, search } = this.state
 		const totalSelected = selectedIds.length
@@ -276,9 +296,11 @@ export default class RecordSelectionList extends Component<
 
 		return (
 			<div className="record-selection__list">
-				<TextContainer>
-					<Text>{`${totalSelected} selected`}</Text>
-				</TextContainer>
+				{showSelectedCount && (
+					<TextContainer>
+						<Text>{`${totalSelected} selected`}</Text>
+					</TextContainer>
+				)}
 
 				{canSearch && (
 					<TextInput
@@ -328,4 +350,8 @@ export default class RecordSelectionList extends Component<
 			</div>
 		)
 	}
+}
+
+RecordSelectionList.defaultProps = {
+	showSelectedCount: true
 }