+++ packages/react-heartwood-components/src/components/Table/Table-story.js
@@ -1,18 +1,26 @@
 // @flow
-import React from 'react'
+import React, { Fragment } from 'react'
 import { storiesOf } from '@storybook/react'
 import { withKnobs, text, boolean } from '@storybook/addon-knobs/react'
 import { generateLocations } from '../../../.storybook/data/tableData'
+import { Formik } from 'formik'
 import Table, { TableSearch, TableFilters } from './index'
 import Layout, { LayoutSection } from '../Layout'
-import Card, { CardHeader } from '../Card'
+import Card, { CardHeader, CardBody } from '../Card'
+import Button from '../Button/Button'
 import Tabs from '../Tabs'
+import {
+	TextInput,
+	FormLayout,
+	FormLayoutGroup,
+	FormLayoutItem
+} from '../Forms'
 
 const stories = storiesOf('Table', module)
 
 stories.addDecorator(withKnobs)
 
-const locations = generateLocations({ amount: 149 })
+const staticLocations = generateLocations({ amount: 149 })
 
 const columns = [
 	{
@@ -44,6 +52,192 @@ const columns = [
 		maxWidth: 500
 	}
 ]
+
+type State = {
+	locations: Array<Object>
+}
+
+type Props = {}
+
+class ExpandableEditableTable extends React.Component<Props, State> {
+	constructor(props) {
+		super(props)
+		this.state = {
+			locations: generateLocations({ amount: 149 })
+		}
+	}
+
+	handleChangeHours = (e, location, dayId) => {
+		const { locations } = this.state
+
+		locations.forEach(l => {
+			if (l.id === location.id) {
+				const updatedSchedule = l.schedule.map(day => {
+					const updatedDay = day
+					if (day.id === dayId) {
+						l.isDirty = true
+						updatedDay.hours = e.target.value
+						updatedDay.isDirty = true
+					}
+					return updatedDay
+				})
+
+				l.schedule = updatedSchedule
+			}
+		})
+
+		this.setState({ locations })
+	}
+
+	handleValidation = async (location: Object, dayId) => {
+		let errors = {}
+
+		const { locations } = this.state
+
+		const updatedLocation = locations.find(l => l.id === location.id)
+		const updatedScheduleDay =
+			updatedLocation && updatedLocation.schedule.find(day => day.id === dayId)
+
+		if (updatedLocation && updatedScheduleDay && !updatedScheduleDay.isDirty) {
+			updatedLocation.isDirty = true
+			updatedScheduleDay.isDirty = true
+			this.setState({ locations })
+		}
+
+		return errors
+	}
+
+	handleSaveHours = async (location: Object, dayId, values) => {
+		const { locations } = this.state
+
+		const updatedLocation = locations.find(l => l.id === location.id)
+		const updatedScheduleDay =
+			updatedLocation && updatedLocation.schedule.find(day => day.id === dayId)
+
+		if (updatedLocation && updatedScheduleDay) {
+			updatedScheduleDay.hours = values.hours || ''
+			updatedLocation.isDirty = false
+			updatedScheduleDay.isDirty = false
+			this.setState({ locations })
+		}
+	}
+
+	renderStoreScheduleForRow = (row: Object) => {
+		const location = this.state.locations[row.index]
+		const schedule = location && location.schedule
+		return schedule ? (
+			<Fragment>
+				<CardHeader title={'Store Schedule'} />
+				<Table
+					isSelectable
+					className="store-schedule-table"
+					data={schedule}
+					columns={[
+						{
+							Header: 'Day',
+							accessor: 'day',
+							minWidth: 100,
+							maxWidth: 300
+						},
+						{
+							Header: 'Hours',
+							accessor: 'hours',
+							minWidth: 100,
+							maxWidth: 500
+						}
+					]}
+					collapseOnDataChange={false}
+					loading={false}
+					pageSize={schedule.length}
+					showPagination={false}
+					subComponentForRow={row => {
+						return (
+							<CardBody>
+								<Formik
+									initialValues={row.original}
+									validate={values =>
+										this.handleValidation(location, row.original.id, values)
+									}
+									render={(formikProps: Object) => {
+										const { handleChange, values } = formikProps
+
+										return (
+											<form onSubmit={formikProps.handleSubmit}>
+												<FormLayout>
+													<FormLayoutGroup>
+														<FormLayoutItem>
+															<TextInput
+																label={`${row.original.day} Store Hours`}
+																onChange={handleChange}
+																name="hours"
+																value={values.hours || ''}
+															/>
+														</FormLayoutItem>
+														<FormLayoutItem spacerTop={true}>
+															<Button
+																kind="primary"
+																onClick={() =>
+																	this.handleSaveHours(
+																		location,
+																		row.original.id,
+																		values
+																	)
+																}
+																disabled={!row.original.isDirty}
+																text="Save"
+															/>
+														</FormLayoutItem>
+													</FormLayoutGroup>
+												</FormLayout>
+											</form>
+										)
+									}}
+								/>
+							</CardBody>
+						)
+					}}
+					rowIsDirty={row => {
+						return row.original.isDirty
+					}}
+					keyField="id"
+				/>
+			</Fragment>
+		) : null
+	}
+
+	render() {
+		const { locations } = this.state
+
+		return (
+			<Table
+				className="store-locations-table"
+				data={locations}
+				columns={columns}
+				sortable={true}
+				defaultPageSize={50}
+				defaultSorted={[
+					{
+						id: 'publicName',
+						desc: false
+					}
+				]}
+				loading={false}
+				collapseOnDataChange={false}
+				paginationProps={{
+					showPages: true,
+					onPageButtonClick: () => console.log('onPageButtonClick'),
+					totalPages: Math.ceil(locations.length / 50),
+					currentPage: 0
+				}}
+				rowIsDirty={row => {
+					return row.original.isDirty
+				}}
+				subComponentForRow={this.renderStoreScheduleForRow}
+				keyField="id"
+			/>
+		)
+	}
+}
 
 stories
 	.add('Table', () => {
@@ -80,7 +274,7 @@ stories
 						)}
 						<Table
 							className="services-table"
-							data={locations}
+							data={staticLocations}
 							columns={columns}
 							sortable={true}
 							defaultPageSize={50}
@@ -94,11 +288,22 @@ stories
 							paginationProps={{
 								showPages: true,
 								onPageButtonClick: () => console.log('onPageButtonClick'),
-								totalPages: Math.ceil(locations.length / 50),
+								totalPages: Math.ceil(staticLocations.length / 50),
 								currentPage: 0
 							}}
 							keyField="id"
 						/>
+					</Card>
+				</LayoutSection>
+			</Layout>
+		)
+	})
+	.add('Expandable Table', () => {
+		return (
+			<Layout width="full-width">
+				<LayoutSection>
+					<Card>
+						<ExpandableEditableTable />
 					</Card>
 				</LayoutSection>
 			</Layout>
@@ -137,7 +342,7 @@ stories
 					)}
 					<Table
 						className="services-table-selectable"
-						data={locations}
+						data={staticLocations}
 						columns={columns}
 						sortable={true}
 						defaultPageSize={50}
@@ -151,7 +356,7 @@ stories
 						paginationProps={{
 							showPages: true,
 							onPageButtonClick: () => console.log('onPageButtonClick'),
-							totalPages: Math.ceil(locations.length / 50),
+							totalPages: Math.ceil(staticLocations.length / 50),
 							currentPage: 0
 						}}
 						isSelectable