+++ packages/react-heartwood-components/src/components/Table/Table.js
@@ -2,10 +2,12 @@
 // Uses React Table. See https://react-table.js.org/#/story/readme details.
 import React, { Component, Fragment } from 'react'
 import ReactTable from 'react-table'
+import { CSSTransition } from 'react-transition-group'
 import cx from 'classnames'
 import { Checkbox } from '../Forms'
-import Pagination from '../Pagination/Pagination'
 import Icon from '../Icon/Icon'
+import Card from '../Card/Card'
+import Pagination from '../Pagination/Pagination'
 import EmptyState from '../EmptyState/EmptyState'
 import ContextMenu from '../ContextMenu/ContextMenu'
 import type { Props as ButtonProps } from '../Button/Button'
@@ -57,7 +59,13 @@ type Props = {
 	noDataSubheadline?: string,
 	noDataPrimaryAction?: PrimaryAction,
 	noDataPrimaryActionButtonKind?: string,
-	noDataPrimaryActionButtonIcon?: string
+	noDataPrimaryActionButtonIcon?: string,
+
+	/** Return a nested sub-component to be added as an expansion level for designated row. */
+	subComponentForRow?: (row: Object) => any,
+
+	/** Called any time row props are updated. Return true if the given row should appear dirty. */
+	rowIsDirty?: (row: Object) => boolean
 }
 
 type PrimaryAction = {
@@ -177,6 +185,8 @@ export default class Table extends Component<Props, State> {
 			noDataPrimaryAction,
 			noDataPrimaryActionButtonKind,
 			noDataPrimaryActionButtonIcon,
+			subComponentForRow,
+			rowIsDirty,
 			...rest
 		} = this.props
 
@@ -266,6 +276,12 @@ export default class Table extends Component<Props, State> {
 				columns={columnsToRender}
 				className={cx('table', className)}
 				sortable={isSelectable && selectedIds.length > 0 ? false : sortable}
+				expanderDefaults={{
+					sortable: false,
+					resizable: false,
+					filterable: false,
+					width: 40
+				}}
 				getTableProps={() => ({
 					className: 'table__inner'
 				})}
@@ -284,13 +300,30 @@ export default class Table extends Component<Props, State> {
 					}),
 					width: 'auto'
 				})}
-				getTrProps={() => ({
-					className: 'table-row',
-					onClick: this.handleClickRow
-				})}
+				getTrGroupProps={(state, rowInfo) => {
+					const expanded = state.expanded[rowInfo.viewIndex]
+					return {
+						className: cx('table-row-group', {
+							'table-row-group--expanded': expanded,
+							'table-row-group--is-dirty': rowInfo.original.isDirty
+						})
+					}
+				}}
+				getTrProps={(state, rowInfo) => {
+					const expanded = state.expanded[rowInfo.viewIndex]
+					rowIsDirty && rowIsDirty(rowInfo)
+					return {
+						className: cx('table-row', {
+							'table-row--expanded': expanded,
+							'table-row--is-dirty': rowInfo.original.isDirty
+						}),
+						onClick: this.handleClickRow
+					}
+				}}
 				getTdProps={(state, rowInfo, column) => ({
 					className: cx('table-cell', {
-						'table-checkbox-cell': column.id === 'checkbox'
+						'table-checkbox-cell': column.id === 'checkbox',
+						'table-expander-cell': column.expander
 					}),
 					width: 'auto'
 				})}
@@ -312,7 +345,29 @@ export default class Table extends Component<Props, State> {
 					primaryActionButtonKind: noDataPrimaryActionButtonKind,
 					primaryActionButtonIcon: noDataPrimaryActionButtonIcon
 				})}
+				SubComponent={
+					subComponentForRow
+						? row => (
+								<CSSTransition
+									in={true}
+									appear={true}
+									classNames="table-subcomponent"
+									timeout={100}
+								>
+									<div className={'table-subcomponent'}>
+										<Card>{subComponentForRow(row)}</Card>
+									</div>
+								</CSSTransition>
+						  )
+						: null
+				}
 				NoDataComponent={EmptyState}
+				ExpanderComponent={
+					<Icon
+						icon={'keyboard_arrow_right'}
+						className={'table-expander-row'}
+					/>
+				}
 				ThComponent={tableProps => {
 					const { toggleSort, className, ...rest } = tableProps
 					const isSortable =
@@ -349,9 +404,7 @@ export default class Table extends Component<Props, State> {
 					)
 				}}
 				PaginationComponent={tableProps =>
-					totalRows &&
-					tableProps.page === 0 &&
-					totalRows <= tableProps.pageSize ? null : (
+					tableProps.page === 0 && totalRows <= tableProps.pageSize ? null : (
 						<div className="table-pagination__wrapper">
 							<Pagination {...paginationProps} {...tableProps} />
 						</div>