+++ packages/react-heartwood-components/src/components/View/View.js
@@ -3,7 +3,6 @@ import React from 'react'
 import cx from 'classnames'
 import HeaderPrimary from '../Core/components/HeaderPrimary/HeaderPrimary'
 import { Sidebar, SidebarFooter } from '../Core'
-import type { Node } from 'react'
 
 type Props = {
 	sidebarItems?: Array<Object>,
@@ -19,7 +18,10 @@ type Props = {
 	isSidebarExpanded?: boolean,
 	isSidebarMobileExpanded?: boolean,
 	onClickSearch?: Function,
-	searchPlaceholder?: string
+	searchPlaceholder?: string,
+
+	/** Menu children (<ListItem> or <li>) */
+	userMenuItems: ReactNode
 }
 
 const View = (props: Props) => {
@@ -37,6 +39,7 @@ const View = (props: Props) => {
 		forceCloseSidebar,
 		onClickSearch,
 		searchPlaceholder,
+		userMenuItems,
 		children
 	} = props
 
@@ -48,19 +51,6 @@ const View = (props: Props) => {
 				'sidebar--is-missing': !sidebarItems || sidebarItems.length === 0
 			})}
 		>
-			{sidebarItems && sidebarItems.length > 0 && (
-				<Sidebar
-					items={sidebarItems}
-					backLink={sidebarBackLink}
-					footer={<SidebarFooter />}
-					isSidebarVisible={isSidebarVisible}
-					isExpanded={isSidebarExpanded}
-					isMobileExpanded={isSidebarMobileExpanded}
-					toggleExpanded={toggleSidebarExpanded}
-					forceCloseSidebar={forceCloseSidebar}
-					side="left"
-				/>
-			)}
 			<HeaderPrimary
 				user={user}
 				organization={organization}
@@ -72,9 +62,29 @@ const View = (props: Props) => {
 				isSidebarVisible={isSidebarMobileExpanded}
 				searchPlaceholder={searchPlaceholder}
 				onClickSearch={onClickSearch}
+				userMenuItems={userMenuItems}
 			/>
 
-			<main className="main-content">{children}</main>
+			<div className="main-content-outer">
+				{sidebarItems && sidebarItems.length > 0 && (
+					<div className="main-content__sidebar">
+						<Sidebar
+							style={{ zIndex: 1 }}
+							items={sidebarItems}
+							backLink={sidebarBackLink}
+							footer={<SidebarFooter />}
+							isSidebarVisible={isSidebarVisible}
+							isExpanded={isSidebarExpanded}
+							isMobileExpanded={isSidebarMobileExpanded}
+							toggleExpanded={toggleSidebarExpanded}
+							forceCloseSidebar={forceCloseSidebar}
+							side="left"
+						/>
+					</div>
+				)}
+
+				<main className="main-content">{children}</main>
+			</div>
 		</div>
 	)
 }